requirements.txt
agora-token-builder==1.0.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
auto-py-to-exe==2.42.0
bidict==0.22.1
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cv2module==0.1.2
cx_Freeze==7.1.0.dev4
cycler==0.12.1
distlib==0.3.7
Django==4.2.6
django-crispy-forms==2.1
django-model-utils==4.4.0
django-widget-tweaks==1.5.0
EasyProcess==1.1
Eel==0.16.0
entrypoint2==1.1
et-xmlfile==1.1.0
fastapi==0.105.0
filelock==3.13.1
Flask==3.0.0
Flask-SocketIO==5.3.6
fonttools==4.50.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.2
gunicorn==21.2.0
h11==0.14.0
idna==3.6
ifaddr==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
macholib==1.16.3
MarkupSafe==2.1.3
matplotlib==3.8.3
modulegraph==0.19.6
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pep517==0.13.1
Pillow==10.1.0
platformdirs==4.0.0
py2app==0.28.7
PyAutoGUI==0.9.54
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
PyGetWindow==0.0.9
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.30
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
python-dateutil==2.9.0.post0
python-engineio==4.8.0
python-socketio==5.10.0
python-xlib==0.33
pytweening==1.0.7
pytz==2024.1
requests==2.31.0
rubicon-objc==0.4.7
setuptools==69.5.1
shiboken6==6.7.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
ssm-python==0.0.2
starlette==0.27.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.24.7
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
whichcraft==0.6.1
winregistry==1.1.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
